Commiting a cookie cutter module sync set of files
This module had pre-existing content which is not the
methodology puppet-openstack community advises.

This commit takes the content from their new-module
documentation and commits it to this repo as a marker
to where the module should have started.

The previous history is being kept for preservation
and will be migrated into upcoming commits to establish
a developmental history based on the cookiecutter and
msync foundation as documented by the puppet openstack
community.

Change-Id: I91def5bf63b88b22153ebefd4018f1252f4b5d2b
Signed-off-by: Dan Radez <dradez@redhat.com>