Add group to policy management
The move of policy.json into code means the file may not exist. We've
added support to ensure that the file exists in the openstacklib but we
need to make sure the permissions are right for each service. This adds
the group information to the policies so it works right.

Depends-On: I26e8b1384f4f69712da9d06a4c565dfd1f17c9ed
Change-Id: Ic0b3bfee864ad58c7469e9eff2265b700ada9a50
Co-Authored-By: Alex Schultz <aschultz@redhat.com>