Disable beaker jobs for now
The centos packaging is broken because congress recently bumped their
psycopg2 requirement to version >= 2.7 but RDO only ships version 2.5.
Basically installing congress doesn't seem to be working on centos or
ubuntu at the moment so we need to kill the beaker jobs so we can land
the stein branch bits.

Change-Id: Iccbf3dbff9f21c1ee2dc075c1324670d66bc0aa7